# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "img2ec",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/skus/{sku_id}/variants/{variant_id}/copy": {
      "get": {
        "tags": [
          "copy"
        ],
        "summary": "List Copy",
        "operationId": "list_copy_api_projects__project_id__skus__sku_id__variants__variant_id__copy_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          },
          {
            "name": "variant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CopyOut"
                  },
                  "title": "Response List Copy Api Projects  Project Id  Skus  Sku Id  Variants  Variant Id  Copy Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/skus/{sku_id}/variants/{variant_id}/copy/regenerate": {
      "post": {
        "tags": [
          "copy"
        ],
        "summary": "Regenerate",
        "operationId": "regenerate_api_projects__project_id__skus__sku_id__variants__variant_id__copy_regenerate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          },
          {
            "name": "variant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Variant Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CopyOut"
                  },
                  "title": "Response Regenerate Api Projects  Project Id  Skus  Sku Id  Variants  Variant Id  Copy Regenerate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CopyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "selling_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selling Points"
          },
          "description_md": {
            "type": "string",
            "title": "Description Md"
          },
          "category_path": {
            "type": "string",
            "title": "Category Path"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags"
          },
          "video_script": {
            "type": "string",
            "title": "Video Script",
            "default": ""
          },
          "detail_template_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail Template Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "platform",
          "title",
          "subtitle",
          "selling_points",
          "description_md",
          "category_path",
          "keywords",
          "hashtags",
          "created_at",
          "updated_at"
        ],
        "title": "CopyOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}